--- src/Microsoft.OData.Edm/Csdl/Parsing/CsdlParserHelpers.cs
@@ -1,22 +1,0 @@
-﻿//---------------------------------------------------------------------
-// <copyright file="CsdlParserHelpers.cs" company="Microsoft">
-//      Copyright (C) Microsoft Corporation. All rights reserved. See License.txt in the project root for license information.
-// </copyright>
-//---------------------------------------------------------------------
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.OData.Edm.Csdl.Json;
-using Microsoft.OData.Edm.Csdl.Parsing.Ast;
-
-namespace Microsoft.OData.Edm.Csdl.Parsing
-{
-    /// <summary>
-    /// Provides functionalities for parsing Schema JSON for Csdl elements.
-    /// </summary>
-    internal static class CsdlParserHelpers
-    {
-
-    }
-}
